refactor: move files to more appropriate places
Also amended some imports with the correct `.js` extension.
