+++ apps/portfolio-frontend/src/components/index.ts
@@ -1,6 +1,2 @@
-export { CountryForm, type CountryFormData } from "./CountryForm.js";
-export { CountryLayout } from "./CountryLayout.js";
 export { DataPageContainer } from "./DataPageContainer.js";
 export { Navigation } from "./Navigation.js";
-export { StockMetadataForm, type StockFormData } from "./StockMetadataForm.js";
-export { StockMetadataLayout } from "./StockMetadataLayout.js";
