+++ apps/portfolio-frontend/src/pages/countries/AddCountryPage.tsx
@@ -2,8 +2,8 @@ import { FormButtonsComponent } from "@codescape-financial/core-ui";
 import { CreateCountryDTO } from "@codescape-financial/portfolio-data-models";
 import { FormEvent, useState } from "react";
 import { useNavigate } from "react-router-dom";
-import { CountryForm, CountryFormData } from "../../components/index.js";
 import { useAxios } from "../../hooks/index.js";
+import { CountryForm, CountryFormData } from "./CountryForm.js";
 
 export const AddCountryPage = () => {
   const navigate = useNavigate();
