+++ apps/portfolio-frontend/src/pages/stocks/StockMetadataForm.tsx
