+++ libs/core-ui/src/lib/components/data-table/index.ts
@@ -1,2 +1,2 @@
-export type { ColumnSchema } from "./ColumnSchema";
-export { DataTable } from "./DataTable";
+export type { ColumnSchema } from "./ColumnSchema.js";
+export { DataTable } from "./DataTable.js";
