+++ libs/core-ui/src/lib/utility/index.ts
@@ -1,1 +1,1 @@
-export { cn } from "./classnames";
+export { cn } from "./classnames.js";
